git-svn-id: https://svn.potix.com/svn/zk1/trunk@2681 dd50bd9b-9913-0410-b9ba-e07a3075be86

+++ zcommon/src/org/zkoss/lang/Threads.java
@@ -86,6 +86,6 @@ public class Threads {
 	 * statments around it.
 	 */
 	public static final void dummy(Object o) {
-		new Integer(Objects.hashCode(o)).intValue();
+		new Integer(Objects.hashCode(o));
 	}
 }
